Removed an unused file and split one into two
