+++ ConcernsCaseWork/ConcernsCaseWork.API.Contracts/Permissions/CasePermissionResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace ConcernsCaseWork.API.Contracts.Permissions;
+
+public class CasePermissionResponse
+{
+	public long CaseId { get; set; }
+	public CasePermission[] Permissions { get; set; }
+}
