+++ ConcernsCaseWork/ConcernsCaseWork.API.Contracts/Permissions/PermissionQueryRequest.cs
@@ -1,11 +1,4 @@
-﻿using Ardalis.GuardClauses;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ConcernsCaseWork.API.Contracts.Permissions
+﻿namespace ConcernsCaseWork.API.Contracts.Permissions
 {
 	public class PermissionQueryRequest
 	{
@@ -19,32 +12,5 @@ namespace ConcernsCaseWork.API.Contracts.Permissions
 		 */
 
 		public long[] CaseIds { get; set; }
-	}
-
-	public class CasePermissionsRequest
-	{
-		public long CaseId { get; set; }
-	}
-
-	public class PermissionQueryResponse
-	{
-		public PermissionQueryResponse()
-		{
-
-		}
-
-		public PermissionQueryResponse(IEnumerable<CasePermissionResponse> casePermissionResponses)
-		{
-			Guard.Against.Null(casePermissionResponses);
-			CasePermissionResponses = casePermissionResponses.ToArray();
-		}
-
-		public CasePermissionResponse[] CasePermissionResponses { get; set; }
-	}
-
-	public class CasePermissionResponse
-	{
-		public long CaseId { get; set; }
-		public CasePermission[] Permissions { get; set; }
 	}
 }
